Extract rich text conversion logic to separate file
